Backport: Properly NULL-terminate GSS receive buffer on error packet reception (#1498)
This pr fixes https://www.postgresql.org/support/security/CVE-2022-41862/ in cloudberry


https://git.postgresql.org/cgit/postgresql.git/commit/?id=71c37797d7bd78266146a5829ab62b3687c47295


Original commit message:

===

pqsecure_open_gss() includes a code path handling error messages with v2-style protocol messages coming from the server.  The client-side buffer holding the error message does not force a NULL-termination, with the data of the server getting copied to the errorMessage of the connection.  Hence, it would be possible for a server to send an unterminated string and copy arbitrary bytes in the buffer receiving the error message in the client, opening the door to a crash or even data exposure.

As at this stage of the authentication process the exchange has not been completed yet, this could be abused by an attacker without Kerberos credentials.  Clients that have a valid kerberos cache are vulnerable as libpq opportunistically requests for it except if gssencmode is disabled.

Author: Jacob Champion
Backpatch-through: 12
Security: CVE-2022-41862